requirements.txt
eventregistry
google-generativeai
